combine fp8 blockwise kernel code, add autotune
