+++ .gitignore
@@ -16,5 +16,3 @@ logs/
 primus_turbo/_build_info.py
 primus_turbo/_version.py
 .coverage*
-tmp_*/
-.rocprofv3/
